% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegChangeR.R
\name{extract_changes_exact}
\alias{extract_changes_exact}
\title{Extract vegetation changes to polygon features}
\usage{
extract_changes_exact(vc_changes, polygons, max_polygons = NULL)
}
\arguments{
\item{vc_changes}{Vegetation changes object from get_VC_optimized0()}

\item{polygons}{SF object with polygon features}

\item{max_polygons}{Maximum number of polygons to process (for testing)}
}
\value{
The input polygons with added columns for mean vegetation change
at each time scale
}
\description{
Calculates net vegetation change statistics for polygon features using
precise extraction with exactextractr.
}
\examples{
\dontrun{
polygons_with_changes <- extract_changes_exact(changes, study_area_polygons)
}

}
