% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegChangeR.R
\name{plot_polygon_changes}
\alias{plot_polygon_changes}
\title{Plot vegetation change distributions}
\usage{
plot_polygon_changes(polygons_with_changes)
}
\arguments{
\item{polygons_with_changes}{SF object from extract_changes_exact()}
}
\value{
ggplot object with the visualization
}
\description{
Creates histogram and density plots of vegetation changes across
different time scales.
}
\examples{
\dontrun{
plot_polygon_changes(polygons_with_changes)
}

}
