% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegChangeR.R
\name{save_changes}
\alias{save_changes}
\title{Save vegetation change results}
\usage{
save_changes(vc_changes, output_dir = "D:/HealthMetric")
}
\arguments{
\item{vc_changes}{Vegetation changes object from analysis functions}

\item{output_dir}{Output directory for saving results}
}
\description{
Exports vegetation change rasters and metadata to specified directory.
}
